worker/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
librosa>=0.10
soundfile>=0.12
matplotlib>=3.8
PyPDF2>=3.0
fonttools>=4.50
scipy>=1.10
python-multipart>=0.0.9
